extracted player funcionality to powerupHelper and weaponhelper, <organisation
